# speed up merge operation for lots of tablets