--- a/core/src/main/java/org/apache/accumulo/core/metadata/schema/TabletMutatorBase.java
+++ b/core/src/main/java/org/apache/accumulo/core/metadata/schema/TabletMutatorBase.java
@@ -18,7 +18,7 @@
  */
 package org.apache.accumulo.core.metadata.schema;
 
-import java.util.Set;
+import java.util.Map;
 
 import org.apache.accumulo.core.client.admin.TabletAvailability;
 import org.apache.accumulo.core.clientImpl.TabletAvailabilityUtil;
@@ -309,15 +309,17 @@ public abstract class TabletMutatorBase<T extends Ample.TabletUpdates<T>>
   }
 
   @Override
-  public T deleteAll(Set<Key> keys) {
+  public T deleteAll(Iterable<Map.Entry<Key,Value>> keys) {
     ByteSequence row = new ArrayByteSequence(mutation.getRow());
-    keys.forEach(key -> {
+    keys.forEach(entry -> {
+      var key = entry.getKey();
       Preconditions.checkArgument(key.getRowData().equals(row), "Unexpected row %s %s", row, key);
       Preconditions.checkArgument(key.getColumnVisibilityData().length() == 0,
           "Non empty column visibility %s", key);
     });
 
-    keys.forEach(key -> {
+    keys.forEach(entry -> {
+      var key = entry.getKey();
       mutation.putDelete(key.getColumnFamily(), key.getColumnQualifier());
     });