Made two changes to speed up merge. First adjusted the sleep time for one of the Fate steps to consider how long a scan took.  For SplitMillionIT this resulted in ReserveTablets.isReady() sleeping for
around 30s instead of 60s.   Second when the merge code was deleting
tablets it was asking TabletMetadata to save key values and those saved
key values were being sorted, even though they did not need to be and
were actually already sorted. Changed code to not sort and the resulted
in DeleteTablets.call() going from 63s to 55s.  Below are the relevant
log messages from before and after runs w/ change to not sort.

```
2024-05-17T18:27:49,635 104 [fate.Fate] DEBUG: Running DeleteTablets.call() FATE:USER:9b3c9abe-a29a-4776-80ee-b0e6c0132d98 took 63328 ms and returned FinishTableRangeOp
```

```
2024-05-17T20:20:31,468 103 [fate.Fate] DEBUG: Running DeleteTablets.call() FATE:USER:3f566052-d88d-44dd-9e8b-ac34ed9ca445 took 54710 ms and returned FinishTableRangeO
```