@cshannon 2024-05-18
If it ends up not being worth making the change in main that is fine too, I just figured I'd bring it up since I noticed it was in main too